use requests, not drest
we're not actually taking advantage of drest's resources abstraction,
and we're going to need to be able to follow arbitrary urls (ie. a link
in a icommons-api response), which drest doesn't like to let you do.
even *with* the resources abstraction, it's not gaining us very much, so
let's toss it.